[FreeBSD] begin converting process_node() to usermode data only.
Rename process_node() to process_vnode() as the former is declared
for all dialects and can't be changed.

Reimplement readvfs() using the statfs() call (whose prototype had
to be re-declared as the _KERNEL define hides it and other things break
without it). Also add the ability to search for filesystems by filename,
which also helps as KERN_VNODE is completely broken.

Use the system-wide xvnode array to find the matching xvnode for
each xfile, and pass it to process_node(). There, make reading the
kernel "struct vnode" optional for now, and start using kinfo_file,
xfile and xvnode fields in favour of the kernel's "struct vnode".